+++ be/src/exprs/es_functions.cpp
@@ -21,7 +21,7 @@ namespace starrocks {
 
 StatusOr<ColumnPtr> ESFunctions::match(FunctionContext* context, const Columns& columns) {
     auto size = columns[0]->size();
-    ColumnBuilder<TYPE_BOOLEAN> result(size);
+    ColumnBuilder<TYPE_BOOLEAN> result(static_cast<int32_t>(size));
     for (int row = 0; row < size; ++row) {
         result.append(true);
     }
